move pyproj function to module geo
