+++ src/nycdb/__init__.py
@@ -4,7 +4,7 @@ __version__ = '0.1.0'
 __author__ = 'ziggy <ziggy@elephant-bird.net>'
 #__all__ = []
 
-from . import typecast, sql, transform, bbl, dof_parser, address
+from . import typecast, sql, transform, bbl, dof_parser, address, geo
 from .database import Database
 from .dataset import Dataset
 from .datasets import datasets
